Move DAQmx error handler to separate file
It doesn't belong in any of the existing files.
