+++ src/Clock.c
@@ -1,6 +1,7 @@
 #include "Clock.h"
 
 #include "DAQConfig.h"
+#include "DAQError.h"
 #include "DeviceImplData.h"
 #include "Waveform.h"
