+++ src/DAQConfig.c
@@ -1,6 +1,7 @@
 #include "DAQConfig.h"
 
 #include "Clock.h"
+#include "DAQError.h"
 #include "Detector.h"
 #include "DeviceImplData.h"
 #include "Scanner.h"
@@ -48,32 +49,6 @@ static bool GetAIPhysChan(OScDev_Device *device, int index, ss8str *chan) {
 
     ss8_destroy(&chans);
     return !notFound;
-}
-
-static char *ErrorCodeDomain() {
-    static char *domainName = NULL;
-    if (domainName == NULL) {
-        domainName = "NI DAQmx";
-        OScDev_Error_RegisterCodeDomain(domainName,
-                                        OScDev_ErrorCodeFormat_I32);
-    }
-    return domainName;
-}
-
-// Must be called immediately after failed DAQmx function
-OScDev_RichError *CreateDAQmxError(int32 nierr) {
-    if (nierr == 0)
-        return OScDev_RichError_OK;
-
-    char buf[1024];
-    DAQmxGetExtendedErrorInfo(buf, sizeof(buf));
-
-    if (nierr > 0) {
-        OScDev_Log_Warning(NULL, buf);
-        return OScDev_RichError_OK;
-    }
-
-    return OScDev_Error_CreateWithCode(ErrorCodeDomain(), nierr, buf);
 }
 
 void SetWaveformParamsFromDevice(OScDev_Device *device,
