+++ src/meson.build
@@ -2,6 +2,7 @@ openscan_nidaq_src = files(
     'Acquisition.c',
     'Clock.c',
     'DAQConfig.c',
+    'DAQError.c',
     'Detector.c',
     'DeviceImplData.c',
     'OpenScanDevice.c',
